pyproject.toml
[project]
name = "point-cloud-labeller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.4",
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.1.0",
    "timm>=1.0.12",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "mobile-sam@git+https://github.com/ChaoningZhang/MobileSAM.git",
]
